+++ hammerdown_server/src/main/java/com/site/hammerdown/common/config/AppConfig.java
@@ -1,0 +1,16 @@
+package com.site.hammerdown.common.config;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableConfigurationProperties(ApplicationProperties.class)
+public class AppConfig {
+
+    @Bean
+    public ObjectMapper modelMapper() {
+        return new ObjectMapper();
+    }
+}
