+++ hammerdown_server/src/main/java/com/site/hammerdown/common/config/ApplicationProperties.java
@@ -1,0 +1,12 @@
+package com.site.hammerdown.common.config;
+
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+
+@ConfigurationProperties(prefix = "hammerdown")
+@Data
+public class ApplicationProperties {
+    private final String name;
+}
