+++ hammerdown_server/src/main/java/com/site/hammerdown/common/model/APIResponseStatus.java
@@ -1,0 +1,6 @@
+package com.site.hammerdown.common.model;
+
+public enum APIResponseStatus {
+    SUCCESS,
+    FAILURE,
+}
