+++ hammerdown_server/src/main/java/com/site/hammerdown/common/model/AppRole.java
@@ -1,0 +1,7 @@
+package com.site.hammerdown.common.model;
+
+public enum AppRole {
+    ROLE_USER,
+    ROLE_ADMIN,
+    ROLE_SUPER_ADMIN
+}
