+++ hammerdown_server/src/main/java/com/site/hammerdown/common/model/Role.java
@@ -1,0 +1,24 @@
+package com.site.hammerdown.common.model;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Role {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer roleId;
+
+    @Enumerated(EnumType.STRING)
+    private AppRole roleName;
+
+    public Role(AppRole appRole) {
+        this.roleName = appRole;
+    }
+
+}
