+++ hammerdown_server/src/main/java/com/site/hammerdown/common/repository/RoleRepository.java
@@ -1,0 +1,14 @@
+package com.site.hammerdown.common.repository;
+
+import com.site.hammerdown.common.model.AppRole;
+import com.site.hammerdown.common.model.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+
+@Repository
+public interface RoleRepository extends JpaRepository<Role, Integer> {
+    Optional<Role> findByRoleName(AppRole appRole);
+}
