--- hammerdown_server/src/main/java/com/site/hammerdown/config/AppConfig.java
@@ -1,16 +1,0 @@
-package com.site.hammerdown.config;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@EnableConfigurationProperties(ApplicationProperties.class)
-public class AppConfig {
-
-    @Bean
-    public ObjectMapper modelMapper() {
-        return new ObjectMapper();
-    }
-}
