--- hammerdown_server/src/main/java/com/site/hammerdown/config/ApplicationProperties.java
@@ -1,12 +1,0 @@
-package com.site.hammerdown.config;
-
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-
-@ConfigurationProperties(prefix = "hammerdown")
-@Data
-public class ApplicationProperties {
-    private final String name;
-}
