--- hammerdown_server/src/main/java/com/site/hammerdown/model/APIResponseStatus.java
@@ -1,6 +1,0 @@
-package com.site.hammerdown.model;
-
-public enum APIResponseStatus {
-    SUCCESS,
-    FAILURE,
-}
