--- hammerdown_server/src/main/java/com/site/hammerdown/model/AppRole.java
@@ -1,7 +1,0 @@
-package com.site.hammerdown.model;
-
-public enum AppRole {
-    ROLE_USER,
-    ROLE_ADMIN,
-    ROLE_SUPER_ADMIN
-}
