--- hammerdown_server/src/main/java/com/site/hammerdown/model/Role.java
@@ -1,24 +1,0 @@
-package com.site.hammerdown.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Role {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer roleId;
-
-    @Enumerated(EnumType.STRING)
-    private AppRole roleName;
-
-    public Role(AppRole appRole) {
-        this.roleName = appRole;
-    }
-
-}
