--- hammerdown_server/src/main/java/com/site/hammerdown/model/User.java
@@ -1,50 +1,0 @@
-package com.site.hammerdown.model;
-
-import jakarta.persistence.*;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.HashSet;
-import java.util.Set;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name ="users",
-        uniqueConstraints = {
-                @UniqueConstraint(columnNames = "userName"),
-                @UniqueConstraint(columnNames = "email")
-        })
-@Builder
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userId;
-
-    @NotBlank
-    private String userName;
-
-    @NotBlank
-    private String password;
-
-    @NotBlank
-    @Email
-    private String email;
-    private String phone;
-
-    private String firstName;
-    private String lastName;
-
-    @ManyToMany(fetch = FetchType.EAGER,
-            cascade = {CascadeType.PERSIST, CascadeType.MERGE})
-    @JoinTable(name = "user_role",
-            joinColumns = @JoinColumn(name = "userId"),
-            inverseJoinColumns = @JoinColumn(name = "roleId"))
-    private Set<Role> roles = new HashSet<>();
-
-}
