--- hammerdown_server/src/main/java/com/site/hammerdown/repository/RoleRepository.java
@@ -1,14 +1,0 @@
-package com.site.hammerdown.repository;
-
-import com.site.hammerdown.model.AppRole;
-import com.site.hammerdown.model.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-
-@Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-    Optional<Role> findByRoleName(AppRole appRole);
-}
