--- hammerdown_server/src/main/java/com/site/hammerdown/repository/UserRepository.java
@@ -1,17 +1,0 @@
-package com.site.hammerdown.repository;
-
-import com.site.hammerdown.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUserName(String username);
-
-    boolean existsByUserName(String username);
-
-    boolean existsByEmail(String email);
-}
