--- hammerdown_server/src/main/java/com/site/hammerdown/security/request/LoginRequest.java
@@ -1,17 +1,0 @@
-package com.site.hammerdown.security.request;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class LoginRequest {
-    @NotBlank
-    private String username;
-
-    @NotBlank
-    private String password;
-}
