+++ hammerdown_server/src/main/java/com/site/hammerdown/user/payload/LoginRequest.java
@@ -1,0 +1,17 @@
+package com.site.hammerdown.user.payload;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class LoginRequest {
+    @NotBlank
+    private String username;
+
+    @NotBlank
+    private String password;
+}
