+++ hammerdown_server/src/main/java/com/site/hammerdown/user/payload/SignupRequest.java
@@ -1,0 +1,31 @@
+package com.site.hammerdown.user.payload;
+
+import com.site.hammerdown.common.model.AppRole;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.Set;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class SignupRequest {
+    @NotBlank
+    @Size(min = 3, max = 20)
+    private String username;
+
+    @NotBlank
+    @Size(max = 50)
+    @Email
+    private String email;
+
+    private Set<AppRole> roles ;
+
+    @NotBlank
+    @Size(min = 6, max = 40)
+    private String password;
+}
