+++ hammerdown_server/src/main/java/com/site/hammerdown/user/service/AuthService.java
@@ -1,0 +1,11 @@
+package com.site.hammerdown.user.service;
+
+import com.site.hammerdown.user.payload.LoginRequest;
+import com.site.hammerdown.user.payload.SignupRequest;
+import com.site.hammerdown.user.payload.UserDTO;
+
+public interface AuthService {
+    UserDTO authenticateUserFromCookie(LoginRequest loginRequest);
+
+    UserDTO registerUser(SignupRequest signUpRequest);
+}
