+++ hammerdown_server/src/main/java/com/site/hammerdown/user/service/UserService.java
@@ -1,0 +1,7 @@
+package com.site.hammerdown.user.service;
+
+import com.site.hammerdown.user.payload.UserDTO;
+
+public interface UserService {
+    UserDTO getUserDetailsFromUserName(String username);
+}
